+++ src/main/java/xatal/sharedz/structures/PublicAbono.java
@@ -1,7 +1,7 @@
 package xatal.sharedz.structures;
 
 public class PublicAbono {
-    public int venta;
+    public Long venta;
     public float cantidad;
     public String fecha;
 }
